+++ flake.nix
@@ -124,6 +124,7 @@
             "sydtest"
             "sydtest-discover"
             "sydtest-aeson"
+            "sydtest-hedgehog"
           ];
         }
         {
@@ -510,6 +511,7 @@
               # sydtest dependencies
               ps.sydtest
               ps.sydtest-discover
+              ps.sydtest-hedgehog
               ps.sydtest-aeson
               ps.validity
               ps.validity-aeson
@@ -606,7 +608,7 @@
 
       # Take a flake app (identified as the key in the 'apps' set), and return a
       # derivation that runs it in the compile phase.
-      # 
+      #
       # In effect, this allows us to run an 'app' as part of the build process (eg: in CI).
       flakeApp2Derivation = system: appName:
         (nixpkgsFor system).runCommand appName { } "${self.apps.${system}.${appName}.program} | tee $out";
@@ -647,7 +649,7 @@
             "ghc810-plutarch:lib:plutarch" = (self.projectMatrix.ghc810.nodev.${system}.flake { }).packages."plutarch:lib:plutarch";
             "ghc810-plutarch:lib:plutarch-test" = (self.projectMatrix.ghc810.nodev.${system}.flake { }).packages."plutarch-test:lib:plutarch-test";
           });
-      # Because `nix flake check` does not work with haskell.nix (due to IFD), 
+      # Because `nix flake check` does not work with haskell.nix (due to IFD),
       # we provide this attribute for running the checks locally, using:
       #   nix build .#check.x86_64-linux
       check = perSystem (system: