+++ plutarch-extra/Plutarch/ListUtils.hs
@@ -1,0 +1,21 @@
+module Plutarch.ListUtils (preverse, pcheckSorted) where
+
+import Plutarch.Prelude
+
+preverse :: (PIsListLike l a) => Term s (l a :--> l a)
+preverse =
+  phoistAcyclic $
+    pfoldl # plam (\ys y -> pcons # y # ys) # pnil
+
+pcheckSorted :: (PIsListLike l a, POrd a) => Term s (l a :--> PBool)
+pcheckSorted =
+  pfix #$ plam $ \self xs ->
+    pelimList
+      ( \x1 xs ->
+          pelimList
+            (\x2 _ -> x1 #<= x2 #&& (self # xs))
+            (pcon PTrue)
+            xs
+      )
+      (pcon PTrue)
+      xs