+++ plutarch-test/plutarch-extra/Plutarch/HelloSpec.hs
@@ -1,8 +1,0 @@
-module Plutarch.HelloSpec (spec) where
-
-import Test.Syd (Spec, describe, it)
-
-spec :: Spec
-spec =
-  describe "Hello-Plutarch-Extra" $
-    it "says Hello Plutarch Extra" $ putStrLn "Hello Plutarch Extra"